deepimg/requirements.txt
streamlit
google-generativeai
requests
pillow
python-dotenv